restructure filesystem. remove photo.
