+++ config/README.md
@@ -1,0 +1,1 @@
+# Face-recognition-sorting
