+++ src/faceMesh.py
@@ -1,0 +1,69 @@
+import cv2
+import numpy as np
+
+# Load YOLOv3-tiny
+net = cv2.dnn.readNet("yolov3-tiny.weights", "yolov3-tiny.cfg")
+layer_names = net.getLayerNames()
+output_layers = [layer_names[i[0] - 1] for i in np.array(net.getUnconnectedOutLayers()).reshape(-1, 1)]
+colors = np.random.uniform(0, 255, size=(4, 3))
+
+# Load COCO classes and find the index for 'person'
+with open("coco.names", "r") as f:
+    classes = [line.strip() for line in f.readlines()]
+face_class_index = classes.index("person")
+
+# Set up the video capture
+cap = cv2.VideoCapture(1)
+cap.set(cv2.CAP_PROP_FRAME_WIDTH, 1280)
+cap.set(cv2.CAP_PROP_FRAME_HEIGHT, 720)
+
+while True:
+    ret, frame = cap.read()
+    height, width, channels = frame.shape
+
+    # Detect faces
+    blob = cv2.dnn.blobFromImage(frame, 0.00392, (416, 416), (0, 0, 0), True, crop=False)
+    net.setInput(blob)
+    outs = net.forward(output_layers)
+
+    # Process the detections
+    class_ids = []
+    confidences = []
+    boxes = []
+
+    for out in outs:
+        for detection in out:
+            scores = detection[5:]
+            class_id = np.argmax(scores)
+            confidence = scores[class_id]
+
+            if class_id == face_class_index and confidence > 0.5:
+                center_x = int(detection[0] * width)
+                center_y = int(detection[1] * height)
+                w = int(detection[2] * width)
+                h = int(detection[3] * height)
+
+                x = int(center_x - w / 2)
+                y = int(center_y - h / 2)
+
+                boxes.append([x, y, w, h])
+                confidences.append(float(confidence))
+                class_ids.append(class_id)
+
+    # Apply non-max suppression
+    indexes = cv2.dnn.NMSBoxes(boxes, confidences, 0.5, 0.4)
+
+    for i in range(min(4, len(boxes))):
+        if i in indexes:
+            x, y, w, h = boxes[i]
+            color = colors[i]
+            cv2.rectangle(frame, (x, y), (x + w, y + h), color, 2)
+            cv2.putText(frame, f"{confidences[i]:.2f}", (x, y + h + 20), cv2.FONT_HERSHEY_SIMPLEX, 0.5, (255, 255, 255), 2)
+
+    cv2.imshow("YOLOv3-tiny Face Detection", frame)
+
+    if cv2.waitKey(1) & 0xFF == ord("q"):
+        break
+    
+cap.release()
+cv2.destroyAllWindows()
